# server/static/openapi.yaml
openapi: 3.0.3
info:
  title: AgriLink API
  description: >
    REST API for AgriLink — a platform connecting farmers with agricultural
    experts through posts, topic-based communities, and direct messaging.
    Authentication uses JWT bearer tokens obtained from /api/auth/login.
  version: 1.0.0
servers:
  - url: https://agrilink-11rw.onrender.com
    description: Production (Render)
  - url: http://localhost:5000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [farmer, expert, admin, user] }
        bio: { type: string, nullable: true }
        location: { type: string, nullable: true }
        profile_image_url: { type: string, nullable: true }
        email_verified: { type: boolean }
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      summary: Register a new account (sends verification email)
      security: []
      responses:
        "201": { description: Created }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/auth/login:
    post:
      tags: [auth]
      summary: Log in and receive a JWT
      security: []
      responses:
        "200": { description: JWT token + user }
        "401": { description: Invalid credentials }
...
  /api/auth/me:
    get:
      tags: [auth]
      summary: Current authenticated user
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/User" } } } }
        "401": { description: Not authenticated }
...
  /api/auth/verify-reset-token/{token}:
    get:
      tags: [auth]
      summary: Check that a reset token is valid
      security: []
      parameters: [{ name: token, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK }, "400": { description: Invalid/expired } }
...